+++ src/account/interface.ts
@@ -5,11 +5,10 @@ import {
   AddTransactionResponse,
   Call,
   DeployContractPayload,
-  Invocation,
   InvocationsDetails,
   Signature,
 } from '../types';
-import { EstimateFee } from '../types/account';
+import { EstimateFee, EstimateFeeDetails } from '../types/account';
 import { BigNumberish } from '../utils/number';
 import { TypedData } from '../utils/typedData/types';
 
@@ -44,7 +43,10 @@ export abstract class AccountInterface extends ProviderInterface {
    *
    * @returns response from addTransaction
    */
-  public abstract estimateFee(invocation: Invocation): Promise<EstimateFee>;
+  public abstract estimateFee(
+    calls: Call | Call[],
+    estimateFeeDetails?: EstimateFeeDetails
+  ): Promise<EstimateFee>;
 
   /**
    * Invoke execute function in account contract