+++ src/types/api.ts
@@ -186,6 +186,14 @@ export type GetContractAddressesResponse = {
   GpsStatementVerifier: string;
 };
 
+export type DeclareTransaction = {
+  type: 'DECLARE';
+  contract_class: CompressedCompiledContract;
+  nonce: BigNumberish;
+  sender_address: BigNumberish;
+  signature: Signature;
+};
+
 export type DeployTransaction = {
   type: 'DEPLOY';
   contract_definition: CompressedCompiledContract;
@@ -205,6 +213,10 @@ export type InvokeFunctionTransaction = {
   max_fee?: BigNumberish;
   version?: BigNumberish;
 };
+
+export interface InvokeFunctionTransactionResponse extends InvokeFunctionTransaction {
+  transaction_hash: string;
+}
 
 export type InvokeFunctionTrace = {
   caller_address: string;
@@ -227,7 +239,7 @@ export type ExecutionResources = {
     bitwise_builtin: number;
     output_builtin: number;
     ecdsa_builtin: number;
-    ec_op_builtin: number;
+    ec_op_builtin?: number;
   };
   n_memory_holes: number;
 };
@@ -237,7 +249,11 @@ export type CallContractTransaction = Omit<
   'type' | 'entry_point_type' | 'nonce'
 >;
 
-export type Transaction = DeployTransaction | InvokeFunctionTransaction;
+export type Transaction = DeclareTransaction | DeployTransaction | InvokeFunctionTransaction;
+export type TransactionResponse =
+  | DeclareTransaction
+  | DeployTransaction
+  | InvokeFunctionTransactionResponse;
 
 export type CallContractResponse = {
   result: string[];
@@ -248,7 +264,7 @@ export type GetBlockResponse = {
   state_root: string;
   block_hash: string;
   transactions: {
-    [txHash: string]: Transaction;
+    [txHash: string]: TransactionResponse;
   };
   timestamp: number;
   transaction_receipts: {
@@ -265,8 +281,9 @@ export type GetBlockResponse = {
       transaction_index: number;
     };
   };
-  previous_block_hash: string;
+  parent_block_hash: string;
   status: Status;
+  gas_price: string;
 };
 
 export type GetCodeResponse = {
@@ -276,9 +293,8 @@ export type GetCodeResponse = {
 
 export type GetTransactionStatusResponse = {
   tx_status: Status;
-  block_hash: string;
+  block_hash?: string;
   tx_failure_reason?: {
-    tx_id: number;
     code: string;
     error_message: string;
   };
@@ -292,7 +308,7 @@ export type GetTransactionTraceResponse = {
     selector: string;
     calldata: RawArgs;
     result: Array<any>;
-    execution_resources: any;
+    execution_resources: ExecutionResources;
     internal_call: Array<any>;
     events: Array<any>;
     messages: Array<any>;
@@ -300,22 +316,33 @@ export type GetTransactionTraceResponse = {
   signature: Signature;
 };
 
-export type GetTransactionResponse = {
+export type SuccessfulTransactionResponse = {
   status: Status;
-  transaction: Transaction;
+  transaction: TransactionResponse;
   block_hash: string;
   block_number: BlockNumber;
   transaction_index: number;
-  transaction_hash: string;
 };
+
+export type FailedTransactionResponse = {
+  status: 'REJECTED';
+  transaction_failure_reason: {
+    code: string;
+    error_message: string;
+  };
+  transaction: TransactionResponse;
+};
+
+export type GetTransactionResponse = SuccessfulTransactionResponse | FailedTransactionResponse;
 
 export type AddTransactionResponse = {
   transaction_hash: string;
   code?: TransactionStatus;
   address?: string;
+  class_hash?: string;
 };
 
-export type TransactionReceiptResponse = {
+export type SuccessfulTransactionReceiptResponse = {
   status: Status;
   transaction_hash: string;
   transaction_index: number;
@@ -323,7 +350,24 @@ export type TransactionReceiptResponse = {
   block_number: BlockNumber;
   l2_to_l1_messages: string[];
   events: string[];
+  actual_fee: string;
+  execution_resources: ExecutionResources;
 };
+
+export type FailedTransactionReceiptResponse = {
+  status: 'REJECTED';
+  transaction_failure_reason: {
+    code: string;
+    error_message: string;
+  };
+  transaction_hash: string;
+  l2_to_l1_messages: string[];
+  events: string[];
+};
+
+export type TransactionReceiptResponse =
+  | SuccessfulTransactionReceiptResponse
+  | FailedTransactionReceiptResponse;
 
 export type EstimateFeeResponse = {
   amount: BN;