+++ CHANGELOG.md
@@ -1,3 +1,83 @@
+## [3.15.3](https://github.com/0xs34n/starknet.js/compare/v3.15.2...v3.15.3) (2022-06-20)
+
+### Bug Fixes
+
+- **dev:** regenerated package-lock.json ([849cb1e](https://github.com/0xs34n/starknet.js/commit/849cb1ea3ffd7ba10b40b232d0ebc46b6599c7ea))
+- use cross-fetch only for jest ([83be37a](https://github.com/0xs34n/starknet.js/commit/83be37a9e3328a44abd9583b8167c3cb8d882790))
+
+## [3.15.2](https://github.com/0xs34n/starknet.js/compare/v3.15.1...v3.15.2) (2022-06-18)
+
+### Bug Fixes
+
+- **cleanup:** cleanup ([a15b6c6](https://github.com/0xs34n/starknet.js/commit/a15b6c6bf13ce7af699c293d697b5136188388c3))
+- **errorcode:** fixed error code >=400 instead of !=200 ([0f16595](https://github.com/0xs34n/starknet.js/commit/0f1659543ba95ce7cac31a5182dad2e33325d4c1))
+- **fetchEndpoint:** error handling and test case fix ([629479f](https://github.com/0xs34n/starknet.js/commit/629479f877aa7d6f39c8d31b2c9449563aadd0e7))
+- **verify:** return false when 500 returned from the gateway ([de3e004](https://github.com/0xs34n/starknet.js/commit/de3e00461730d6fa112046169470d7a603baa296))
+
+## [3.15.1](https://github.com/0xs34n/starknet.js/compare/v3.15.0...v3.15.1) (2022-06-17)
+
+### Bug Fixes
+
+- response type for get-transaction ([a232725](https://github.com/0xs34n/starknet.js/commit/a232725099893b45fd6feda0a4cac7d05f903435))
+
+# [3.15.0](https://github.com/0xs34n/starknet.js/compare/v3.14.1...v3.15.0) (2022-06-16)
+
+### Bug Fixes
+
+- **tests:** update test ([c71f482](https://github.com/0xs34n/starknet.js/commit/c71f482facdaac914c45f4bf91f48e05a930abff))
+
+### Features
+
+- add calculateContractAddressFromHash ([e22c346](https://github.com/0xs34n/starknet.js/commit/e22c3464036f97eee3c617d3790aac35b3d95379))
+
+## [3.14.1](https://github.com/0xs34n/starknet.js/compare/v3.14.0...v3.14.1) (2022-06-15)
+
+### Bug Fixes
+
+- update api typings ([44796af](https://github.com/0xs34n/starknet.js/commit/44796af4849b6bab3d99065bb1e1948e4ea0b55e))
+
+# [3.14.0](https://github.com/0xs34n/starknet.js/compare/v3.13.1...v3.14.0) (2022-06-15)
+
+### Bug Fixes
+
+- remove redundant \_abi from declareContract() ([53d6578](https://github.com/0xs34n/starknet.js/commit/53d6578b932ed6046b5e0df83d748673d7efc3d5))
+
+### Features
+
+- support contract declaration API introduced at SN 0.9.0 ([ca6203f](https://github.com/0xs34n/starknet.js/commit/ca6203f93471d7fb421d580e07d6de7c183e40f3))
+
+## [3.13.1](https://github.com/0xs34n/starknet.js/compare/v3.13.0...v3.13.1) (2022-06-14)
+
+### Bug Fixes
+
+- account estimate fee interface ([51e8cf6](https://github.com/0xs34n/starknet.js/commit/51e8cf6858acd5aed003f8f1c4d755088684f134))
+
+# [3.13.0](https://github.com/0xs34n/starknet.js/compare/v3.12.3...v3.13.0) (2022-06-14)
+
+### Bug Fixes
+
+- account tests ([8e86c0e](https://github.com/0xs34n/starknet.js/commit/8e86c0e813924fa5cf8f708acd708f9bea1bb7c3))
+- add TEST_RPC_URL env var ([8972e0b](https://github.com/0xs34n/starknet.js/commit/8972e0bc4b4180a6f28b7fce736db76b28426a4d))
+- change getStorageAt key param type to BigNumberish ([b8a8389](https://github.com/0xs34n/starknet.js/commit/b8a8389d64c3f845be8302054cbc579b4b3d3fb9))
+- contract tests ([f14fe11](https://github.com/0xs34n/starknet.js/commit/f14fe117504634451540f50ce780230716f539dc))
+- deps versions ([68f8373](https://github.com/0xs34n/starknet.js/commit/68f837390e3a4cba56e134f57b88966818ea94ff))
+- env format ([665ab0b](https://github.com/0xs34n/starknet.js/commit/665ab0b0cc725f83d05628f53ee00b63f2801979))
+- lockfile ([839bc15](https://github.com/0xs34n/starknet.js/commit/839bc1519744ea6df044b4929ec219e632e50db5))
+- move env up ([405aca2](https://github.com/0xs34n/starknet.js/commit/405aca2eb07c882e25435c1b18ee5018e0970c6c))
+- move private key to env ([2f4d22c](https://github.com/0xs34n/starknet.js/commit/2f4d22c47b6442479a355b4ee117ba1a5995420f))
+- remove accountContract test ([d239dab](https://github.com/0xs34n/starknet.js/commit/d239dab6bb2872897501e6fc90b2c41650c2ff8f))
+- remove maxFee: 0 ([9569819](https://github.com/0xs34n/starknet.js/commit/9569819c3ca36e8005b6d591d206888f5689f867))
+- secrets ([d080436](https://github.com/0xs34n/starknet.js/commit/d080436acdd13571192cc19292e8ba0661caebc2))
+- test ([3ca4bd3](https://github.com/0xs34n/starknet.js/commit/3ca4bd37c9edd03ca5be09ad2e9daf837d833f7c))
+- tests ([593dfa1](https://github.com/0xs34n/starknet.js/commit/593dfa1da130e7e0ef1aa6edd971bbec771d16b8))
+
+### Features
+
+- add test account to fixtures ([3a1150c](https://github.com/0xs34n/starknet.js/commit/3a1150cfdb44053a18591e38eb30d112115864be))
+- getTestProvider util ([8d6f0b9](https://github.com/0xs34n/starknet.js/commit/8d6f0b96738ee5ba48ab95356441964d8b8ef6c4))
+- update account guide ([b5275b3](https://github.com/0xs34n/starknet.js/commit/b5275b313897e9a6fd4b1e52f850aa5fb189cb23))
+- update erc20 guide ([416da13](https://github.com/0xs34n/starknet.js/commit/416da13b3690463897dfae318135d6f46198c5ec))
+
 ## [3.12.3](https://github.com/0xs34n/starknet.js/compare/v3.12.2...v3.12.3) (2022-05-30)
 
 ### Bug Fixes